Move out plugins and init package. Remove samples image.
