+++ docker/Dockerfile.adapters-py
@@ -1,13 +1,13 @@
 FROM python:3.9.13-slim
 
 ENV PYTHONUNBUFFERED=1
-ARG APP_PATH=/opt/app
-ENV APP_PATH=$APP_PATH
-ENV PYTHONPATH=$APP_PATH
-WORKDIR $APP_PATH
+ARG PROJECT_PATH=/opt/savant
+ENV PROJECT_PATH=$PROJECT_PATH
+ENV PYTHONPATH=$PROJECT_PATH
+WORKDIR $PROJECT_PATH
 
-COPY adapters/requirements.txt ${APP_PATH}/adapters/requirements.txt
-RUN python -m pip install --no-cache-dir -r ${APP_PATH}/adapters/requirements.txt
+COPY adapters/requirements.txt $PROJECT_PATH/adapters/requirements.txt
+RUN python -m pip install --no-cache-dir -r $PROJECT_PATH/adapters/requirements.txt
 
-COPY savant ${APP_PATH}/savant
-COPY adapters/python ${APP_PATH}/adapters/python
+COPY savant $PROJECT_PATH/savant
+COPY adapters/python $PROJECT_PATH/adapters/python
