+++ gst_plugins/python/logger.py
