+++ gst_plugins/python/zeromq_properties.py
