--- requirements/samples.txt
@@ -1,1 +1,0 @@
-scipy~=1.10
