+++ scripts/run_sink.py
@@ -150,7 +150,7 @@ def display_sink(
         zmq_type=in_type,
         zmq_bind=in_bind,
         sync=sync,
-        entrypoint='/opt/app/adapters/ds/sinks/display.sh',
+        entrypoint='/opt/savant/adapters/ds/sinks/display.sh',
         envs=envs,
         volumes=[f'{xsock}:{xsock}', f'{xauth}:{xauth}'],
         with_gpu=True,
@@ -202,7 +202,7 @@ def meta_json_sink(
         zmq_endpoint=in_endpoint,
         zmq_type=in_type,
         zmq_bind=in_bind,
-        entrypoint='/opt/app/adapters/python/sinks/metadata_json.py',
+        entrypoint='/opt/savant/adapters/python/sinks/metadata_json.py',
         envs=envs,
         volumes=[f'{target_dir}:{target_dir}'],
         docker_image=docker_image,
@@ -251,7 +251,7 @@ def image_files_sink(
         zmq_endpoint=in_endpoint,
         zmq_type=in_type,
         zmq_bind=in_bind,
-        entrypoint='/opt/app/adapters/python/sinks/image_files.py',
+        entrypoint='/opt/savant/adapters/python/sinks/image_files.py',
         envs=envs,
         volumes=[f'{target_dir}:{target_dir}'],
         docker_image=docker_image,
@@ -293,7 +293,7 @@ def video_files_sink(
         zmq_endpoint=in_endpoint,
         zmq_type=in_type,
         zmq_bind=in_bind,
-        entrypoint='/opt/app/adapters/gst/sinks/video_files.sh',
+        entrypoint='/opt/savant/adapters/gst/sinks/video_files.sh',
         envs=envs,
         volumes=[f'{location}:{location}'],
         docker_image=docker_image,
